Review unit tests Where the class is now in the folder "messageHandling", move the tests to match And adjust namespaces in the unit tests project
