+++ JustSaying.AwsTools.UnitTests/MessageHandling/SqsNotificationListener/BaseQueuePollingTest.cs
@@ -1,0 +1,89 @@
+using System;
+using System.Collections.Generic;
+using System.Threading;
+using System.Threading.Tasks;
+using Amazon;
+using Amazon.SQS;
+using Amazon.SQS.Model;
+using JustBehave;
+using JustSaying.AwsTools.MessageHandling;
+using JustSaying.Messaging.MessageHandling;
+using JustSaying.Messaging.MessageSerialisation;
+using JustSaying.Messaging.Monitoring;
+using JustSaying.TestingFramework;
+using NSubstitute;
+
+namespace JustSaying.AwsTools.UnitTests.MessageHandling.SqsNotificationListener
+{
+    public abstract class BaseQueuePollingTest : BehaviourTest<JustSaying.AwsTools.MessageHandling.SqsNotificationListener>
+    {
+        protected const string QueueUrl = "url";
+        protected IAmazonSQS Sqs;
+        protected GenericMessage DeserialisedMessage;
+        protected const string MessageBody = "object";
+        protected IHandler<GenericMessage> Handler;
+        protected IMessageMonitor Monitor;
+        protected IMessageSerialisationRegister SerialisationRegister;
+        protected IMessageLock MessageLock;
+        protected readonly string _messageTypeString = typeof(GenericMessage).ToString();
+
+        protected override JustSaying.AwsTools.MessageHandling.SqsNotificationListener CreateSystemUnderTest()
+        {
+            return new JustSaying.AwsTools.MessageHandling.SqsNotificationListener(new SqsQueueByUrl(RegionEndpoint.EUWest1, QueueUrl, Sqs), SerialisationRegister, Monitor, null, MessageLock);
+        }
+
+        protected override void Given()
+        {
+            Sqs = Substitute.For<IAmazonSQS>();
+            SerialisationRegister = Substitute.For<IMessageSerialisationRegister>();
+            Monitor = Substitute.For<IMessageMonitor>();
+            Handler = Substitute.For<IHandler<GenericMessage>>();
+
+            var response = GenerateResponseMessage(_messageTypeString, Guid.NewGuid());
+            
+            Sqs.ReceiveMessageAsync(
+                    Arg.Any<ReceiveMessageRequest>(), 
+                    Arg.Any<CancellationToken>())
+                .Returns(
+                    x => Task.FromResult(response),
+                    x => Task.FromResult(new ReceiveMessageResponse()));
+
+            DeserialisedMessage = new GenericMessage { RaisingComponent = "Component" };
+            SerialisationRegister.DeserializeMessage(Arg.Any<string>()).Returns(DeserialisedMessage);
+        }
+
+        protected override void When()
+        {
+            SystemUnderTest.AddMessageHandler(() => Handler);
+            SystemUnderTest.Listen();
+        }
+
+        protected ReceiveMessageResponse GenerateResponseMessage(string messageType, Guid messageId)
+        {
+            return new ReceiveMessageResponse
+            {
+                Messages = new List<Message> {       
+                    new Message
+                    {   
+                        MessageId = messageId.ToString(),
+                        Body = SqsMessageBody(messageType)
+                    },
+                    new Message
+                    {
+                        MessageId = messageId.ToString(),
+                        Body = "{\"Subject\":\"SOME_UNKNOWN_MESSAGE\"," + "\"Message\":\"SOME_RANDOM_MESSAGE\"}"
+                    }}
+            };
+        }
+
+        protected string SqsMessageBody(string messageType)
+        {
+            return "{\"Subject\":\"" + messageType + "\"," + "\"Message\":\"" + MessageBody + "\"}";
+        }
+
+        public override void PostAssertTeardown()
+        {
+            SystemUnderTest.StopListening();
+        }
+    }
+}
