+++ JustSaying.AwsTools.UnitTests/MessageHandling/SqsNotificationListener/WhenMessageHandlingFails.cs
@@ -1,0 +1,26 @@
+using System.Threading.Tasks;
+using Amazon.SQS.Model;
+using JustBehave;
+using JustSaying.TestingFramework;
+using NSubstitute;
+
+namespace JustSaying.AwsTools.UnitTests.MessageHandling.SqsNotificationListener
+{
+    public class WhenMessageHandlingFails : BaseQueuePollingTest
+    {
+        protected override void Given()
+        {
+            base.Given();
+            Handler.Handle(Arg.Any<GenericMessage>()).ReturnsForAnyArgs(false);
+        }
+
+        [Then]
+        public async Task FailedMessageIsNotRemovedFromQueue()
+        {
+            // The un-handled one is however.
+            await Patiently.VerifyExpectationAsync(
+                () => Sqs.DidNotReceiveWithAnyArgs().DeleteMessage(
+                        Arg.Any<DeleteMessageRequest>()));
+        }
+    }
+}
